feat: edit module class for CosmicCorner, using location to switch them.
